+++ tests/test-expandable-dashboard-recent-comments.php
@@ -6,6 +6,10 @@ class unittest_c2c_ExpandableDashboardRecentComments extends c2c_ExpandableDashb
 	public static function is_comment_initially_expanded( $comment ) {
 		return parent::is_comment_initially_expanded( $comment );
 	}
+
+	public static function is_text_excerpted( $text ) {
+		return parent::is_text_excerpted( $text );
+	}
 }
 
 
@@ -129,4 +133,10 @@ class Expandable_Dashboard_Recent_Comments_Test extends WP_UnitTestCase {
 
 		$this->assertTrue( unittest_c2c_ExpandableDashboardRecentComments::is_comment_initially_expanded( $comment ) );
 	}
+
+	public function tests_is_text_excerpted() {
+		$this->assertTrue( unittest_c2c_ExpandableDashboardRecentComments::is_text_excerpted( 'This is excerpted&hellip;' ) );
+		$this->assertFalse( unittest_c2c_ExpandableDashboardRecentComments::is_text_excerpted( 'This is not excerpted.' ) );
+	}
+
 }